ResumeScoring/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0